refactor : move test utils to test/utils
