+++ packages/vhd-lib/src/Vhd/VhdAbstract.integ.spec.js
@@ -7,7 +7,7 @@ import { getSyncedHandler } from '@xen-orchestra/fs'
 import { Disposable, pFromCallback } from 'promise-toolbox'
 
 import { openVhd } from '../index'
-import { createRandomFile, convertFromRawToVhd, createRandomVhdDirectory } from '../_testUtils'
+import { createRandomFile, convertFromRawToVhd, createRandomVhdDirectory } from '../tests/utils'
 import { VhdAbstract } from './VhdAbstract'
 
 let tempDir = null
