+++ packages/vhd-lib/src/tests/utils.js
@@ -48,3 +48,40 @@ export async function recoverRawContent(vhdName, rawName, originalSize) {
     await execa('truncate', ['-s', originalSize, rawName])
   }
 }
+
+export async function createRandomVhdDirectory(path, sizeMB) {
+  fs.mkdir(path)
+  const rawFileName = `${path}/temp.raw`
+  await createRandomFile(rawFileName, sizeMB)
+  const vhdFileName = `${path}/vhd`
+  await convertFromRawToVhd(rawFileName, vhdFileName)
+
+  const srcVhd = await fs.open(vhdFileName, 'r')
+
+  const footer = Buffer.alloc(512)
+  await fs.read(srcVhd, footer, 0, footer.length, 0)
+  await fs.writeFile(path + '/footer', footer)
+
+  const header = Buffer.alloc(1024)
+  await fs.read(srcVhd, header, 0, header.length, 512)
+  await fs.writeFile(path + '/header', header)
+
+  await fs.close(srcVhd)
+
+  // a BAT , with at most 512 blocks of 2MB
+  const bat = Buffer.alloc(512, 1)
+  await fs.writeFile(path + '/bat', bat)
+
+  // copy blocks
+  const srcRaw = await fs.open(rawFileName, 'r')
+  const blockDataSize = 512 * 4096
+  const bitmap = Buffer.alloc(4096)
+  await fs.mkdir(path + '/blocks/')
+  await fs.mkdir(path + '/blocks/1/')
+  for (let i = 0, offset = 0; i < sizeMB; i++, offset += blockDataSize) {
+    const blockData = Buffer.alloc(blockDataSize)
+    await fs.read(srcRaw, blockData, offset)
+    await fs.writeFile(path + '/blocks/1/' + i, Buffer.concat([bitmap, blockData]))
+  }
+  await fs.close(srcRaw)
+}
